make shebang portable
Also this does not allow run this script on operating systems such
as *BSD that install bash to /usr/local/bin,
it allows to run it on linux distributions such as NixOS/Guix.

See https://github.com/systemd/systemd/pull/5816 on why this is a good
idea, which has been agreed on also by other Linux distro maintainers.

Also this would break if bash was not in PATH this is highly unlikely
because than also all the other tools that are installed in similar
directories would be missing.